✅ Phase G.3完了: VisualFocusManager.js分割 - Issue #103
- Main Controller Pattern適用によりファイルサイズを2,520語→1,264語に削減（50%削減）
- 4つのコンポーネントに分割：
  * FocusStateManager.js (470語) - フォーカス状態管理
  * FocusEffectRenderer.js (610語) - 視覚効果描画
  * FocusEventHandler.js (600語) - イベント処理
  * FocusAccessibilitySupport.js (796語) - アクセシビリティサポート
- WCAG 2.1 AA準拠とキーボードナビゲーション完全保持
- 既存API完全互換、後方互換性100%維持
